k_means update 1
I filled out the functions in k_means and added functions to create a test set to work with. There are still many many bugs, but In thought I would push this so that everybody can see the progress. If anyone has come up with better ways to perform these calculations, feel free to overwrite them